+++ src/ferricstore/__init__.py
@@ -102,7 +102,7 @@ QueueWorker = QueueFlowWorker
 QueueWorkerResult = QueueFlowWorkerResult
 AsyncQueueWorker = AsyncQueueFlowWorker
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 __all__ = [
     "ApprovalResult",